Greetings I pushed another update rn.
After having a short testing on the app I noticed the bug hidden in Stock Outbound and Return Delete Button Where

when you add 2 items for outbound then edit one of the item's quantity and remove it will not update the record this might be the one causing the issue.

Please let me know if you still encountered the issue.

I will do quick scan on this on my another free time.